create context file that allows booting of API as a subprocess
