+++ src/indigobot/__main__.py
@@ -3,104 +3,37 @@ This is the main chatbot program/file for conversational capabilities and info d
 """
 
 import readline  # Required for using arrow keys in CLI
-from typing import Sequence
+import subprocess
+import os
+import time
 
-from langchain.chains import create_history_aware_retriever, create_retrieval_chain
-from langchain.chains.combine_documents import create_stuff_documents_chain
-from langchain_core.messages import AIMessage, BaseMessage, HumanMessage
-from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
-from langgraph.checkpoint.memory import MemorySaver
-from langgraph.graph import START, StateGraph
-from langgraph.graph.message import add_messages
-
-from typing_extensions import Annotated, TypedDict
-
-from indigobot.config import llms, vectorstores
-from indigobot.utils import custom_loader
-
-llm = llms["gpt"]
+from indigobot.utils.custom_loader import start_loader
+from indigobot.context import chatbot_app
+from indigobot.config import CURRENT_DIR
 
 
-class ChatState(TypedDict):
-    """
-    A dictionary that represents the state of a chat interaction/history.
-    """
-
-    input: str
-    chat_history: Annotated[Sequence[BaseMessage], add_messages]
-    context: str
-    answer: str
+def load():
+    load_res = input("Would you like to execute the loader? (y/n) ")
+    if load_res == "y":
+        try:
+            start_loader()
+        except Exception as e:
+            print(f"Error booting loader: {e}")
 
 
-def call_model(state: ChatState):
-    """
-    Call the model with the given state and return the response.
-
-    :param state: The state containing input, chat history, context, and answer.
-    :type state: ChatState
-    :return: Updated state with chat history, context, and answer.
-    :rtype: dict
-    """
-    response = rag_chain.invoke(state)
-    return {
-        "chat_history": [
-            HumanMessage(state["input"]),
-            AIMessage(response["answer"]),
-        ],
-        "context": response["context"],
-        "answer": response["answer"],
-    }
+def api():
+    load_res = input("Would you like to enable the API? (y/n) ")
+    if load_res == "y":
+        try:
+            api_path = os.path.join(CURRENT_DIR, "quick_api.py")
+            process = subprocess.Popen(["python", api_path])
+            # process.wait()
+            time.sleep(10)
+        except Exception as e:
+            print(f"Error booting api: {e}")
 
 
-retriever = vectorstores["gpt"].as_retriever()
-
-### Contextualize question ###
-contextualize_q_system_prompt = (
-    "Reformulate the user's question into a standalone question, "
-    "considering the chat history. Return the original question if no reformulation needed."
-)
-contextualize_q_prompt = ChatPromptTemplate.from_messages(
-    [
-        ("system", contextualize_q_system_prompt),
-        MessagesPlaceholder("chat_history"),
-        ("human", "{input}"),
-    ]
-)
-history_aware_retriever = create_history_aware_retriever(
-    llm, retriever, contextualize_q_prompt
-)
-
-### Answer question ###
-system_prompt = (
-    "You are an assistant that answers questions/provides information about "
-    "social services in Portland, Oregon. Use the following pieces of "
-    "retrieved context to answer the question. If you don't know the answer, "
-    "say that you don't know. Use three sentences maximum and keep the answer concise."
-    "\n\n"
-    "{context}"
-)
-qa_prompt = ChatPromptTemplate.from_messages(
-    [
-        ("system", system_prompt),
-        MessagesPlaceholder("chat_history"),
-        ("human", "{input}"),
-    ]
-)
-question_answer_chain = create_stuff_documents_chain(llm, qa_prompt)
-rag_chain = create_retrieval_chain(history_aware_retriever, question_answer_chain)
-
-workflow = StateGraph(state_schema=ChatState)
-workflow.add_edge(START, "model")
-workflow.add_node("model", call_model)
-
-memory = MemorySaver()
-chatbot_app = workflow.compile(checkpointer=memory)
-
-# Configuration constants
-thread_config = {"configurable": {"thread_id": "abc123"}}
-
-
-def main(skip_loader: bool = False) -> None:
+def main(skip_loader: bool = False, skip_api: bool = False) -> None:
     """
     Main function that runs the interactive chat loop.
     Handles user input and displays model responses.
@@ -110,9 +43,13 @@ def main(skip_loader: bool = False) -> None:
         skip_loader (bool): If True, skips the loader prompt. Useful for testing.
     """
     if not skip_loader:
-        load_res = input("Would you like to execute the loader? (y/n) ")
-        if load_res == "y":
-            custom_loader.main()
+        load()
+
+    if not skip_api:
+        api()
+
+    # Configuration constants
+    thread_config = {"configurable": {"thread_id": "abc123"}}
 
     while True:
         try:
@@ -128,12 +65,12 @@ def main(skip_loader: bool = False) -> None:
             else:
                 break
         except Exception as e:
-            print(e)
+            print(f"Error with llm input: {e}")
 
 
 if __name__ == "__main__":
     try:
-        main(skip_loader=False)
+        main(skip_loader=False, skip_api=False)
     except KeyboardInterrupt:
         print("\nExiting...")
     except Exception as e:
